# backend/docs/swagger.yaml
definitions:
  domain.ActivateModelReq:
    properties:
      model_id:
        type: string
    required:
    - model_id
    type: object
...
  domain.CheckModelReq:
    properties:
      api_header:
        type: string
      api_key:
        type: string
      api_version:
        description: for azure openai
        type: string
      base_url:
        type: string
      model:
        type: string
      provider:
        allOf:
        - $ref: '#/definitions/domain.ModelProvider'
        enum:
        - OpenAI
        - Ollama
        - DeepSeek
        - SiliconFlow
        - Moonshot
        - Other
        - AzureOpenAI
    required:
    - base_url
    - model
    - provider
    type: object
  domain.CheckModelResp:
    properties:
      content:
        type: string
      error:
        type: string
    type: object
...
  domain.CreateKnowledgeBaseReq:
    properties:
      name:
        type: string
    required:
    - name
    type: object
  domain.CreateModelReq:
    properties:
      api_header:
        type: string
      api_key:
        type: string
      api_version:
        description: for azure openai
        type: string
      base_url:
        type: string
      model:
        type: string
      provider:
        allOf:
        - $ref: '#/definitions/domain.ModelProvider'
        enum:
        - OpenAI
        - Ollama
        - DeepSeek
        - SiliconFlow
        - Moonshot
        - Other
        - AzureOpenAI
    required:
    - base_url
    - model
    - provider
    type: object
...
  domain.GetProviderModelListResp:
    properties:
      models:
        items:
          $ref: '#/definitions/domain.ProviderModelListItem'
        type: array
    type: object
...
  domain.KnowledgeBaseDetail:
    properties:
      id:
        type: string
      name:
        type: string
    type: object
...
  domain.ModelDetailResp:
    properties:
      api_header:
        type: string
      api_key:
        type: string
      api_version:
        description: for azure openai
        type: string
      base_url:
        type: string
      completion_tokens:
        type: integer
      created_at:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      model:
        type: string
      prompt_tokens:
        type: integer
      provider:
        $ref: '#/definitions/domain.ModelProvider'
      total_tokens:
        type: integer
      updated_at:
        type: string
    type: object
...
  domain.ModelProvider:
    enum:
    - OpenAI
    - Ollama
    - DeepSeek
    - Moonshot
    - SiliconFlow
    - AzureOpenAI
    - Other
    type: string
    x-enum-varnames:
    - ModelProviderBrandOpenAI
    - ModelProviderBrandOllama
    - ModelProviderBrandDeepSeek
    - ModelProviderBrandMoonshot
    - ModelProviderBrandSiliconFlow
    - ModelProviderBrandAzureOpenAI
    - ModelProviderBrandOther
...
  domain.ProviderModelListItem:
    properties:
      model:
        type: string
    type: object
...
  domain.Response:
    properties:
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
...
  domain.UpdateKnowledgeBaseReq:
    properties:
      id:
        type: string
      name:
        type: string
    required:
    - id
    - name
    type: object
  domain.UpdateModelReq:
    properties:
      api_header:
        type: string
      api_key:
        type: string
      api_version:
        description: for azure openai
        type: string
      base_url:
        type: string
      id:
        type: string
      model:
        type: string
    required:
    - base_url
    - id
    - model
    type: object
...
info:
  contact: {}
paths:
...
  /api/v1/knowledge_base:
    post:
      consumes:
      - application/json
      description: CreateKnowledgeBase
      parameters:
      - description: CreateKnowledgeBase Request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.CreateKnowledgeBaseReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Response'
      summary: CreateKnowledgeBase
      tags:
      - knowledge_base
  /api/v1/knowledge_base/detail:
    delete:
      consumes:
      - application/json
      description: DeleteKnowledgeBase
      parameters:
      - description: Knowledge Base ID
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Response'
      summary: DeleteKnowledgeBase
      tags:
      - knowledge_base
    get:
      consumes:
      - application/json
      description: GetKnowledgeBaseDetail
      parameters:
      - description: Knowledge Base ID
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/domain.Response'
            - properties:
                data:
                  $ref: '#/definitions/domain.KnowledgeBaseDetail'
              type: object
      summary: GetKnowledgeBaseDetail
      tags:
      - knowledge_base
    put:
      consumes:
      - application/json
      description: UpdateKnowledgeBase
      parameters:
      - description: UpdateKnowledgeBase Request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.UpdateKnowledgeBaseReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Response'
      summary: UpdateKnowledgeBase
      tags:
      - knowledge_base
...
  /api/v1/model:
    delete:
      consumes:
      - application/json
      description: delete model
      parameters:
      - description: model id
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Response'
      tags:
      - model
    post:
      consumes:
      - application/json
      description: create model
      parameters:
      - description: create model request
        in: body
        name: model
        required: true
        schema:
          $ref: '#/definitions/domain.CreateModelReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Response'
      summary: create model
      tags:
      - model
    put:
      consumes:
      - application/json
      description: update model
      parameters:
      - description: update model request
        in: body
        name: model
        required: true
        schema:
          $ref: '#/definitions/domain.UpdateModelReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Response'
      tags:
      - model
  /api/v1/model/activate:
    post:
      consumes:
      - application/json
      description: activate model
      parameters:
      - description: activate model request
        in: body
        name: model
        required: true
        schema:
          $ref: '#/definitions/domain.ActivateModelReq'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Response'
      summary: activate model
      tags:
      - model
  /api/v1/model/check:
    post:
      consumes:
      - application/json
      description: check model
      parameters:
      - description: check model request
        in: body
        name: model
        required: true
        schema:
          $ref: '#/definitions/domain.CheckModelReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/domain.Response'
            - properties:
                data:
                  $ref: '#/definitions/domain.CheckModelResp'
              type: object
      summary: check model
      tags:
      - model
  /api/v1/model/detail:
    get:
      consumes:
      - application/json
      description: get model detail
      parameters:
      - description: model id
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/domain.Response'
            - properties:
                data:
                  $ref: '#/definitions/domain.ModelDetailResp'
              type: object
      summary: get model detail
      tags:
      - model
...
  /api/v1/model/provider/supported:
    get:
      consumes:
      - application/json
      description: get provider supported model list
      parameters:
      - in: query
        name: api_header
        type: string
      - in: query
        name: api_key
        type: string
      - in: query
        name: base_url
        required: true
        type: string
      - enum:
        - SiliconFlow
        - OpenAI
        - Ollama
        - DeepSeek
        - Moonshot
        - AzureOpenAI
        in: query
        name: provider
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/domain.Response'
            - properties:
                data:
                  $ref: '#/definitions/domain.GetProviderModelListResp'
              type: object
      summary: get provider supported model list
      tags:
      - model
...
swagger: "2.0"